requirements.txt
pyyaml
loguru
pytest
mkdocs-material
black
